fix(line3) renamed fromData to fromPointAndDirection
