+++ src/core/math/line3/clone.test.js
@@ -1,22 +1,22 @@
 const test = require('ava')
-const { create, clone, fromData, toString } = require('./index')
+const { create, clone, fromPointAndDirection, toString } = require('./index')
 
 const { compareVectors } = require('../../../../test/helpers/index')
 
 test('line3: clone() should return a new line3 with same values', (t) => {
-  const org1 = fromData([0, 0, 0], [1, 0, 0])
+  const org1 = fromPointAndDirection([0, 0, 0], [1, 0, 0])
   const obs1 = clone(org1)
   t.true(compareVectors(obs1[0], [0, 0, 0]))
   t.true(compareVectors(obs1[1], [1, 0, 0]))
   t.not(obs1, org1)
 
-  const org2 = fromData([1, 2, 3], [1, 0, 1])
+  const org2 = fromPointAndDirection([1, 2, 3], [1, 0, 1])
   const obs2 = clone(org2)
   t.true(compareVectors(obs2[0], [1, 2, 3]))
   t.true(compareVectors(obs2[1], [0.7071067690849304, 0, 0.7071067690849304]))
   t.not(obs2, org2)
 
-  const org3 = fromData([-1, -2, -3], [0, -1, -1])
+  const org3 = fromPointAndDirection([-1, -2, -3], [0, -1, -1])
   const obs3 = clone(org3)
   t.true(compareVectors(obs3[0], [-1, -2, -3]))
   t.true(compareVectors(obs3[1], [0, -0.7071067690849304, -0.7071067690849304]))
@@ -25,7 +25,7 @@ test('line3: clone() should return a new line3 with same values', (t) => {
 
 test('line3: clone() with two params should update a line3 with same values', (t) => {
   const line1 = create()
-  const org1 = fromData([0, 0, 0], [1, 0, 0])
+  const org1 = fromPointAndDirection([0, 0, 0], [1, 0, 0])
   const ret1 = clone(line1, org1)
   t.true(compareVectors(line1[0], [0, 0, 0]))
   t.true(compareVectors(line1[1], [1, 0, 0]))
@@ -34,7 +34,7 @@ test('line3: clone() with two params should update a line3 with same values', (t
   t.not(ret1, org1)
 
   const line2 = create()
-  const org2 = fromData([1, 2, 3], [1, 0, 1])
+  const org2 = fromPointAndDirection([1, 2, 3], [1, 0, 1])
   const ret2 = clone(line2, org2)
   t.true(compareVectors(line2[0], [1, 2, 3]))
   t.true(compareVectors(line2[1], [0.7071067690849304, 0, 0.7071067690849304]))
@@ -43,7 +43,7 @@ test('line3: clone() with two params should update a line3 with same values', (t
   t.not(ret2, org2)
 
   const line3 = create()
-  const org3 = fromData([-1, -2, -3], [0, -1, -1])
+  const org3 = fromPointAndDirection([-1, -2, -3], [0, -1, -1])
   const ret3 = clone(line3, org3)
   t.true(compareVectors(line3[0], [-1, -2, -3]))
   t.true(compareVectors(line3[1], [0, -0.7071067690849304, -0.7071067690849304]))
