+++ src/core/math/line3/create.js
@@ -1,6 +1,6 @@
 const vec3 = require('../vec3')
 
-const fromData = require('./fromData')
+const fromPointAndDirection = require('./fromPointAndDirection')
 
 /**
  * Create an unbounded 3D line, centered at 0,0,0 and lying on X axis.
@@ -10,7 +10,7 @@ const fromData = require('./fromData')
 const create = () => {
   let point = vec3.fromValues(0, 0, 0)
   let direction = vec3.fromValues(0, 0, 1)
-  return fromData(point, direction)
+  return fromPointAndDirection(point, direction)
 }
 
 module.exports = create
