+++ src/core/math/line3/fromPointAndDirection.js
